Extract user data in a separate file
