+++ frontend/test/metabase/scenarios/auth/signin.cy.spec.js
@@ -1,4 +1,5 @@
-import { browse, restore, signIn, signOut, USERS } from "__support__/cypress";
+import { browse, restore, signIn, signOut } from "__support__/cypress";
+import { USERS } from "__support__/cypress_data";
 
 const sizes = [[1280, 800], [640, 360]];
 const { admin } = USERS;
