+++ frontend/test/metabase/scenarios/setup/user_settings.cy.spec.js
@@ -1,5 +1,6 @@
 // Migrated from frontend/test/metabase/user/UserSettings.integ.spec.js
-import { restore, signInAsNormalUser, USERS } from "__support__/cypress";
+import { restore, signInAsNormalUser } from "__support__/cypress";
+import { USERS } from "__support__/cypress_data";
 const { first_name, last_name, email } = USERS.normal;
 
 const requestsCount = alias =>
